+++ src/extension.ts
@@ -4,20 +4,26 @@ import * as vscode from 'vscode';
 import * as config from './config';
 import * as util from './util';
 import * as diagnostics from './diagnostics';
+import * as format from './format';
 
 export async function activate(context: vscode.ExtensionContext) {
-    const rootPaths = await util.getRootPaths();
+    const rootPaths = await util.findRootPaths();
 
     if (rootPaths.length === 0) {
-        vscode.window.showWarningMessage('Rust Assist: No `Cargo.toml` files were found in the workspace.');
+        vscode.window.showWarningMessage('Rust Assist: No `Cargo.toml` files were found in the workspace, unable to start plugin.');
+        return;
     }
+
+    // ====================================================
+    // Diagnostics
+    // ====================================================
 
     const rustDiagnostics = vscode.languages.createDiagnosticCollection("rust");
     const diagnosticManager = new diagnostics.DiagnosticsManager(rootPaths, rustDiagnostics);
 
     context.subscriptions.push(
         vscode.commands.registerCommand('rust-assist.refreshDiagnostics', async () => {
-            await diagnosticManager.refreshAll();
+            diagnosticManager.refreshAll();
         })
     );
 
@@ -25,13 +31,29 @@ export async function activate(context: vscode.ExtensionContext) {
         context.subscriptions.push(
             vscode.workspace.onDidSaveTextDocument(async (document: vscode.TextDocument) => {
                 if (document.languageId === 'rust') {
-                    await diagnosticManager.refreshAll();
+                    diagnosticManager.refreshAll();
                 }
             })
         );
     }
 
     if (config.diagnosticsOnStartup()) {
-        await diagnosticManager.refreshAll();
+        diagnosticManager.refreshAll();
+    }
+
+    // ====================================================
+    // Formatting
+    // ====================================================
+
+    const formatManager = new format.FormatManager(rootPaths, config.formatMode());
+
+    if (config.formatOnSave()) {
+        context.subscriptions.push(
+            vscode.workspace.onDidSaveTextDocument(async (document: vscode.TextDocument) => {
+                if (document.languageId === 'rust') {
+                    formatManager.format(document.uri.fsPath);
+                }
+            })
+        );
     }
 }