+++ src/util.ts
@@ -25,10 +25,24 @@ export function spawn(
     });
 }
 
-export async function getRootPaths(): Promise<Array<string>> {
+/**
+ * Find all paths in the workspace that contain a Cargo.toml file.
+ * 
+ * @returns An array of paths.
+ */
+export async function findRootPaths(): Promise<Array<string>> {
     let roots: Array<string> = [];
     (await vscode.workspace.findFiles('**/Cargo.toml')).forEach(path => {
         roots.push(path.fsPath.replace(/[/\\]?Cargo\.toml$/, ''));
     });
     return roots;
+}
+
+export function findMatchingRoot(roots: Array<string>, key: string) {
+    for (const root of roots) {
+        if (key.startsWith(root)) {
+            return root;
+        }
+    }
+    return undefined;
 }