+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### Pending
+- Make configuration names more descriptive
+
 ### 0.1.2 - 2018-05-23
 - Fix issue with warnings not being reissued for lib projects
 - Update the logo