+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-### Pending
+### 0.2.0 - 2018-05-23
 - Make configuration names more descriptive
+- Add formatting support through `rustfmt`
 
 ### 0.1.2 - 2018-05-23
 - Fix issue with warnings not being reissued for lib projects